requirement.txt
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4
scrapy
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
djangorestframework
docopt==0.6.2
idna==2.9
jsonschema==3.2.0
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.3.0
Django
pyrsistent==0.15.7
psycopg2
gunicorn
PyYAML==5.3
requests==2.23.0
six==1.14.0
texttable==1.6.2
urllib3==1.25.8
websocket-client==0.57.0
peeweee